requirements.txt
boto==2.49.0
boto3==1.9.127
botocore==1.12.127
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-extensions==2.1.6
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-request-logging==0.5.4
django-rest-swagger==2.1.2
django-smtp-ssl==1.0
django-storages==1.7.1
djangorestframework==3.7.3
docutils==0.14
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
Markdown==2.6.9
MarkupSafe==1.0
mysql-connector-python==2.0.4
mysqlclient==1.4.2
oauthlib==2.0.6
openapi-codec==1.3.2
protobuf==3.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
simplejson==3.12.0
six==1.11.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wfastcgi==3.0.0
whitenoise==4.1.2
